SFIO sfvscanf: squash -Wsign-conversion warnings for 'n_str'

+++ lib/sfio/sfvscanf.c
@@ -185,11 +185,11 @@ int sfvscanf(FILE *f, const char *form, va_list args)
 				goto t_arg;
 			    if ((t_str = argv.s) &&
 				(n_str = (int) ft->size) < 0)
-				n_str = strlen(t_str);
+				n_str = (ssize_t)strlen(t_str);
 			} else {
 			  t_arg:
 			    if ((t_str = va_arg(args, char *)))
-				 n_str = strlen(t_str);
+				 n_str = (ssize_t)strlen(t_str);
 			}
 		    }
 		    goto loop_flags;
